+++ test/Microsoft.DotNet.GenAPI.Tests/CSharpFileBuilderTests.cs
@@ -44,8 +44,8 @@ namespace Microsoft.DotNet.GenAPI.Tests
                 new ConsoleLog(MessageImportance.Low),
                 stringWriter,
                 loader,
-                SymbolFilterFactory.GetSymbolFilterFromList([], includeInternalSymbols, includeEffectivelyPrivateSymbols, includeExplicitInterfaceImplementationSymbols),
-                SymbolFilterFactory.GetAttributeFilterFromList(excludedAttributeList, includeInternalSymbols, includeEffectivelyPrivateSymbols, includeExplicitInterfaceImplementationSymbols),
+                CompositeSymbolFilter.GetSymbolFilterFromList([], includeInternalSymbols, includeEffectivelyPrivateSymbols, includeExplicitInterfaceImplementationSymbols),
+                CompositeSymbolFilter.GetAttributeFilterFromList(excludedAttributeList, includeInternalSymbols, includeEffectivelyPrivateSymbols, includeExplicitInterfaceImplementationSymbols),
                 header: string.Empty,
                 exceptionMessage: null,
                 includeAssemblyAttributes: false,
